Read Seabird .hex files

The Seabird hex format is not publicly defined anywhere, but is a mix of binary integer representations scaled in various ways to represent voltage, frequency, count, and NMEA output. These functions are best considered as entertainment only.

read_sbe_hex_lines(file, skip = 0, n_max = Inf, header = read_sbe_header(file))

Arguments

file
A file, URL, or connection. Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed; URLs will be automatically downloaded. See readr::read_lines() for a full description of how this parameter is interpreted.

skip
Number of non-header rows to skip.

n_max
Maximum number of rows to read.

header
A previously read value obtained from read_sbe_header().